style(test): :art: hardware modelのテストコードを分割
